STRWEB/LR1/LR5/СosmetologyCenter/requirements.txt
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0